Altirra 2.00

+++ src/h/vd2/Riza/audioformat.h
@@ -50,10 +50,10 @@ namespace nsVDWinFormats {
 	struct WaveFormatExPCM : public WaveFormatEx{
 		WaveFormatExPCM(uint32 samplingRate, uint32 channels, uint32 precision) {
 			mFormatTag = kWAVE_FORMAT_PCM;
-			mChannels = channels;
+			mChannels = (uint16)channels;
 			mSamplesPerSec = samplingRate;
-			mBitsPerSample = precision;
-			mBlockAlign = mChannels * ((precision + 7) >> 3);
+			mBitsPerSample = (uint16)precision;
+			mBlockAlign = (uint16)(mChannels * ((precision + 7) >> 3));
 			mAvgBytesPerSec = samplingRate * mBlockAlign;
 			mSize = 0;
 		}
